Change the test Data to random copy/insert/delete modifications.
Make RandIter.skip() correctly skip past the cycle wrap point.

Make alltests() use the bsize for both the data and chunker, and return tsize
and bsize along with the results. This means the the avg index in the results
is a multiple of the bsize, and result data files also include the data and
chunker bsize used for the test.

In graphs.py use the bsize from the data file instead of assuming it's 1024
bytes. Remove the KB references for avg sizes, since it's actualy multiples of
the bsize.