Change test runs and graphs for better perfomance visualisation.
Only run tests needed for the graphs we want. Increase resolution of min
values tested. Only run full min range for max=8.0 and full max range for
min=0.0, and add specific tests for values we want to graph in the middle.

For graphs.py fix typo in FileName(). Add PerfVsMinLimitByAlg() graph, and add
"ByAvg" to PerfVsMinLimit() and PerfVsMaxLimit() names. Use "%.1f" for
formating min limit now that we use 0.0, 0.1, ...0.7 for the limit range.
Generate less but more meaningful graphs.