Add RC4Chunker and extra graphs.
This adds the "Regression Chunker" idea from a microsoft paper. It improves
handling of max-len truncations by "rewinding" back to an earlier point with a
weaker hash judgement criteria to use as the break point.

This required adding getstate() setstate() support to RandIter, Data, Chunker
and stats1.Sample classes so that the scan state could be saved and "rewound" back. It
also meant moving the data scanning from runtest() to a Chunker.scan() method.

In runtest() we add some additional runs to better analyse the impact of
min_len for very small max_len.

In graphs we add PerfVsMinLimitByAvg(), SizeAvgVsAvgTarget(),
PerfVsMinLimitByAlg(), and PerfVsAvgSize() for the smallest max_len, and
PerfVsMaxLimitByAvg() for min_len=0.5. We drop the PerfVsAvgSize for
min_len=0.2, max_len=4.0.